+++ apps/code-examples/src/app/code-examples/ag-grid/data-entry-grid/basic/data-entry-grid-docs-demo.component.ts
@@ -15,14 +15,14 @@ import {
   ValueFormatterParams,
 } from 'ag-grid-community';
 
-import { DataEntryGridContextMenuComponent } from './data-entry-grid-demo-context-menu.component';
-import { AG_GRID_DEMO_DATA } from './data-entry-grid-demo-data';
-import { DataEntryGridEditModalContext } from './data-entry-grid-demo-edit-modal-context';
-import { DataEntryGridEditModalComponent } from './data-entry-grid-demo-edit-modal.component';
+import { DataEntryGridContextMenuComponent } from './data-entry-grid-docs-demo-context-menu.component';
+import { AG_GRID_DEMO_DATA } from './data-entry-grid-docs-demo-data';
+import { DataEntryGridEditModalContext } from './data-entry-grid-docs-demo-edit-modal-context';
+import { DataEntryGridEditModalComponent } from './data-entry-grid-docs-demo-edit-modal.component';
 
 @Component({
   selector: 'app-data-entry-grid-demo',
-  templateUrl: './data-entry-grid-demo.component.html',
+  templateUrl: './data-entry-grid-docs-demo.component.html',
   changeDetection: ChangeDetectionStrategy.OnPush,
 })
 export class DataEntryGridDemoComponent {
