+++ apps/code-examples/src/app/code-examples/ag-grid/data-entry-grid/basic/data-entry-grid-docs-demo.module.ts
@@ -5,7 +5,7 @@ import { SkySearchModule } from '@skyux/lookup';
 
 import { AgGridModule } from 'ag-grid-angular';
 
-import { DataEntryGridDemoComponent } from './data-entry-grid-demo.component';
+import { DataEntryGridDemoComponent } from './data-entry-grid-docs-demo.component';
 
 @NgModule({
   declarations: [DataEntryGridDemoComponent],
