+++ apps/code-examples/src/app/code-examples/ag-grid/data-entry-grid/data-manager-added/data-manager-data-entry-grid-docs-demo-edit-modal.component.ts
@@ -22,24 +22,24 @@ import {
   GridReadyEvent,
   ICellEditorParams,
   IRowNode,
-  NewValueParams,
 } from 'ag-grid-community';
 
 import {
   AgGridDemoRow,
   DEPARTMENTS,
   JOB_TITLES,
-} from './data-entry-grid-demo-data';
-import { DataEntryGridEditModalContext } from './data-entry-grid-demo-edit-modal-context';
+} from './data-manager-data-entry-grid-docs-demo-data';
+import { DataEntryGridEditModalContext } from './data-manager-data-entry-grid-docs-demo-edit-modal-context';
 
 @Component({
   standalone: true,
-  selector: 'app-data-entry-grid-demo-edit-modal',
-  templateUrl: './data-entry-grid-demo-edit-modal.component.html',
+  selector: 'app-data-manager-data-entry-grid-demo-edit-modal',
+  templateUrl:
+    './data-manager-data-entry-grid-docs-demo-edit-modal.component.html',
   changeDetection: ChangeDetectionStrategy.OnPush,
   imports: [AgGridModule, SkyAgGridModule, SkyModalModule],
 })
-export class DataEntryGridEditModalComponent {
+export class DataManagerDataEntryGridEditModalComponent {
   protected gridData: AgGridDemoRow[];
   protected gridOptions: GridOptions;
 
@@ -47,8 +47,8 @@ export class DataEntryGridEditModalComponent {
   #gridApi: GridApi | undefined;
 
   protected readonly instance = inject(SkyModalInstance);
-  readonly #agGridSvc = inject(SkyAgGridService);
-  readonly #changeDetectorRef = inject(ChangeDetectorRef);
+  readonly #agGridService = inject(SkyAgGridService);
+  readonly #changeDetector = inject(ChangeDetectorRef);
   readonly #context = inject(DataEntryGridEditModalContext);
 
   constructor() {
@@ -56,36 +56,12 @@ export class DataEntryGridEditModalComponent {
       {
         field: 'name',
         headerName: 'Name',
-        type: SkyCellType.Text,
-        cellRendererParams: {
-          skyComponentProperties: {
-            validator: (value: string): boolean => String(value).length <= 10,
-            validatorMessage: `Value exceeds maximum length`,
-          },
-        },
-        cellEditorParams: {
-          skyComponentProperties: {
-            maxlength: 10,
-          },
-        },
-        editable: true,
       },
       {
         field: 'age',
         headerName: 'Age',
         type: SkyCellType.Number,
-        cellRendererParams: {
-          skyComponentProperties: {
-            validator: (value: number): boolean => value >= 18,
-            validatorMessage: `Age must be 18+`,
-          },
-        },
         maxWidth: 60,
-        cellEditorParams: {
-          skyComponentProperties: {
-            min: 18,
-          },
-        },
         editable: true,
       },
       {
@@ -102,7 +78,11 @@ export class DataEntryGridEditModalComponent {
         cellEditorParams: (
           params: ICellEditorParams
         ): { skyComponentProperties: SkyAgGridDatepickerProperties } => {
-          return { skyComponentProperties: { minDate: params.data.startDate } };
+          return {
+            skyComponentProperties: {
+              minDate: params.data.startDate,
+            },
+          };
         },
       },
       {
@@ -116,15 +96,13 @@ export class DataEntryGridEditModalComponent {
           return {
             skyComponentProperties: {
               data: DEPARTMENTS,
-              selectionChange: (
-                change: SkyAutocompleteSelectionChange
-              ): void => {
+              selectionChange: (change): void => {
                 this.#departmentSelectionChange(change, params.node);
               },
             },
           };
         },
-        onCellValueChanged: (event: NewValueParams): void => {
+        onCellValueChanged: (event): void => {
           if (event.newValue !== event.oldValue) {
             this.#clearJobTitle(event.node);
           }
@@ -171,7 +149,7 @@ export class DataEntryGridEditModalComponent {
           skyComponentProperties: {
             validator: (value: Date): boolean =>
               !!value && value > new Date(1985, 9, 26),
-            validatorMessage: 'Please enter a future date',
+            validatorMessage: 'Enter a future date.',
           },
         },
         editable: true,
@@ -180,22 +158,22 @@ export class DataEntryGridEditModalComponent {
 
     this.gridData = this.#context.gridData;
 
-    this.gridOptions = {
+    const gridOptions: GridOptions = {
       columnDefs: this.#columnDefs,
       onGridReady: (gridReadyEvent): void => this.onGridReady(gridReadyEvent),
     };
 
-    this.gridOptions = this.#agGridSvc.getEditableGridOptions({
-      gridOptions: this.gridOptions,
+    this.gridOptions = this.#agGridService.getEditableGridOptions({
+      gridOptions,
     });
 
-    this.#changeDetectorRef.markForCheck();
+    this.#changeDetector.markForCheck();
   }
 
   public onGridReady(gridReadyEvent: GridReadyEvent): void {
     this.#gridApi = gridReadyEvent.api;
     this.#gridApi.sizeColumnsToFit();
-    this.#changeDetectorRef.markForCheck();
+    this.#changeDetector.markForCheck();
   }
 
   #departmentSelectionChange(
@@ -210,11 +188,14 @@ export class DataEntryGridEditModalComponent {
   #clearJobTitle(node: IRowNode | null): void {
     if (node) {
       node.data.jobTitle = undefined;
-      this.#changeDetectorRef.markForCheck();
 
       if (this.#gridApi) {
-        this.#gridApi.refreshCells({ rowNodes: [node] });
+        this.#gridApi.refreshCells({
+          rowNodes: [node],
+        });
       }
     }
+
+    this.#changeDetector.markForCheck();
   }
 }
