+++ apps/code-examples/src/app/code-examples/ag-grid/data-entry-grid/data-manager-added/data-manager-data-entry-grid-docs-demo-filter-modal.component.ts
