+++ apps/code-examples/src/app/code-examples/ag-grid/data-entry-grid/data-manager-added/data-manager-data-entry-grid-docs-demo-view-grid.component.ts
@@ -25,11 +25,12 @@ import {
 } from 'ag-grid-community';
 import { Subject, takeUntil } from 'rxjs';
 
-import { AgGridDemoRow } from './data-manager-data-entry-grid-demo-data';
+import { AgGridDemoRow } from './data-manager-data-entry-grid-docs-demo-data';
 
 @Component({
   selector: 'app-data-manager-data-entry-grid-demo-view-grid',
-  templateUrl: './data-manager-data-entry-grid-demo-view-grid.component.html',
+  templateUrl:
+    './data-manager-data-entry-grid-docs-demo-view-grid.component.html',
   changeDetection: ChangeDetectionStrategy.OnPush,
 })
 export class DataManagerDataEntryGridDemoViewGridComponent
