+++ apps/code-examples/src/app/code-examples/ag-grid/data-entry-grid/inline-help/data-entry-grid-docs-demo-edit-modal-context.ts
@@ -1,4 +1,4 @@
-import { AgGridDemoRow } from './data-manager-data-entry-grid-demo-data';
+import { AgGridDemoRow } from './data-entry-grid-docs-demo-data';
 
 export class DataEntryGridEditModalContext {
   public gridData: AgGridDemoRow[] = [];
