+++ apps/code-examples/src/app/code-examples/ag-grid/data-grid/basic-multiselect/basic-multiselect-data-grid-docs-demo.module.ts
@@ -3,7 +3,7 @@ import { SkyAgGridModule } from '@skyux/ag-grid';
 
 import { AgGridModule } from 'ag-grid-angular';
 
-import { BasicMultiselectDataGridDemoComponent } from './basic-multiselect-data-grid-demo.component';
+import { BasicMultiselectDataGridDemoComponent } from './basic-multiselect-data-grid-docs-demo.component';
 
 @NgModule({
   declarations: [BasicMultiselectDataGridDemoComponent],
