+++ apps/code-examples/src/app/code-examples/ag-grid/data-grid/data-manager-multiselect/data-manager-multiselect-data-grid-docs-demo-filter-modal.component.ts
@@ -17,7 +17,7 @@ import { SkyModalInstance, SkyModalModule } from '@skyux/modals';
   standalone: true,
   selector: 'app-data-manager-multiselect-data-grid-demo-filter-modal',
   templateUrl:
-    './data-manager-multiselect-data-grid-demo-filter-modal.component.html',
+    './data-manager-multiselect-data-grid-docs-demo-filter-modal.component.html',
   changeDetection: ChangeDetectionStrategy.OnPush,
   imports: [FormsModule, SkyCheckboxModule, SkyIdModule, SkyModalModule],
 })
