+++ apps/code-examples/src/app/code-examples/ag-grid/data-grid/data-manager-multiselect/data-manager-multiselect-data-grid-docs-demo-view-grid.component.html
