+++ apps/code-examples/src/app/code-examples/ag-grid/data-grid/data-manager/data-manager-data-grid-docs-demo-data.ts
