+++ apps/code-examples/src/app/features/ag-grid.module.ts
@@ -1,20 +1,20 @@
 import { NgModule } from '@angular/core';
 import { RouterModule, Routes } from '@angular/router';
 
-import { DataEntryGridDemoComponent as DataEntryGridBasicDataEntryGridDemoComponent } from '../code-examples/ag-grid/data-entry-grid/basic/data-entry-grid-demo.component';
-import { DataEntryGridDemoModule as DataEntryGridBasicDataEntryGridDemoModule } from '../code-examples/ag-grid/data-entry-grid/basic/data-entry-grid-demo.module';
-import { DataManagerDataEntryGridDemoComponent as DataEntryGridDataManagerAddedDataManagerDataEntryGridDemoComponent } from '../code-examples/ag-grid/data-entry-grid/data-manager-added/data-manager-data-entry-grid-demo.component';
-import { DataManagerDataEntryGridDemoModule as DataEntryGridDataManagerAddedDataManagerDataEntryGridDemoModule } from '../code-examples/ag-grid/data-entry-grid/data-manager-added/data-manager-data-entry-grid-demo.module';
-import { DataEntryGridDemoComponent as DataEntryGridInlineHelpDataEntryGridDemoComponent } from '../code-examples/ag-grid/data-entry-grid/inline-help/data-entry-grid-demo.component';
-import { DataEntryGridDemoModule as DataEntryGridInlineHelpDataEntryGridDemoModule } from '../code-examples/ag-grid/data-entry-grid/inline-help/data-entry-grid-demo.module';
-import { BasicMultiselectDataGridDemoComponent as DataGridBasicMultiselectDataGridDemoComponent } from '../code-examples/ag-grid/data-grid/basic-multiselect/basic-multiselect-data-grid-demo.component';
-import { BasicMultiselectDataGridDemoModule } from '../code-examples/ag-grid/data-grid/basic-multiselect/basic-multiselect-data-grid-demo.module';
-import { BasicDataGridDemoComponent as DataGridBasicDataGridDemoComponent } from '../code-examples/ag-grid/data-grid/basic/basic-data-grid-demo.component';
-import { BasicDataGridDemoModule as DataGridBasicBasicDataGridDemoModule } from '../code-examples/ag-grid/data-grid/basic/basic-data-grid-demo.module';
-import { DataManagerMultiselectDataGridDemoComponent as DataGridDataManagerAddedDataManagerMultiselectDataGridDemoComponent } from '../code-examples/ag-grid/data-grid/data-manager-multiselect/data-manager-multiselect-data-grid-demo.component';
-import { DataManagerMultiselectDataGridDemoModule } from '../code-examples/ag-grid/data-grid/data-manager-multiselect/data-manager-multiselect-data-grid-demo.module';
-import { DataManagerDataGridDemoComponent as DataGridDataManagerAddedDataManagerDataGridDemoComponent } from '../code-examples/ag-grid/data-grid/data-manager/data-manager-data-grid-demo.component';
-import { DataManagerDataGridDemoModule } from '../code-examples/ag-grid/data-grid/data-manager/data-manager-data-grid-demo.module';
+import { DataEntryGridDemoComponent as DataEntryGridBasicDataEntryGridDemoComponent } from '../code-examples/ag-grid/data-entry-grid/basic/data-entry-grid-docs-demo.component';
+import { DataEntryGridDemoModule as DataEntryGridBasicDataEntryGridDemoModule } from '../code-examples/ag-grid/data-entry-grid/basic/data-entry-grid-docs-demo.module';
+import { DataManagerDataEntryGridDemoComponent as DataEntryGridDataManagerAddedDataManagerDataEntryGridDemoComponent } from '../code-examples/ag-grid/data-entry-grid/data-manager-added/data-manager-data-entry-grid-docs-demo.component';
+import { DataManagerDataEntryGridDemoModule as DataEntryGridDataManagerAddedDataManagerDataEntryGridDemoModule } from '../code-examples/ag-grid/data-entry-grid/data-manager-added/data-manager-data-entry-grid-docs-demo.module';
+import { DataEntryGridDemoComponent as DataEntryGridInlineHelpDataEntryGridDemoComponent } from '../code-examples/ag-grid/data-entry-grid/inline-help/data-entry-grid-docs-demo.component';
+import { DataEntryGridDemoModule as DataEntryGridInlineHelpDataEntryGridDemoModule } from '../code-examples/ag-grid/data-entry-grid/inline-help/data-entry-grid-docs-demo.module';
+import { BasicMultiselectDataGridDemoComponent as DataGridBasicMultiselectDataGridDemoComponent } from '../code-examples/ag-grid/data-grid/basic-multiselect/basic-multiselect-data-grid-docs-demo.component';
+import { BasicMultiselectDataGridDemoModule } from '../code-examples/ag-grid/data-grid/basic-multiselect/basic-multiselect-data-grid-docs-demo.module';
+import { BasicDataGridDemoComponent as DataGridBasicDataGridDemoComponent } from '../code-examples/ag-grid/data-grid/basic/basic-data-grid-docs-demo.component';
+import { BasicDataGridDemoModule as DataGridBasicBasicDataGridDemoModule } from '../code-examples/ag-grid/data-grid/basic/basic-data-grid-docs-demo.module';
+import { DataManagerMultiselectDataGridDemoComponent as DataGridDataManagerAddedDataManagerMultiselectDataGridDemoComponent } from '../code-examples/ag-grid/data-grid/data-manager-multiselect/data-manager-multiselect-data-grid-docs-demo.component';
+import { DataManagerMultiselectDataGridDemoModule } from '../code-examples/ag-grid/data-grid/data-manager-multiselect/data-manager-multiselect-data-grid-docs-demo.module';
+import { DataManagerDataGridDemoComponent as DataGridDataManagerAddedDataManagerDataGridDemoComponent } from '../code-examples/ag-grid/data-grid/data-manager/data-manager-data-grid-docs-demo.component';
+import { DataManagerDataGridDemoModule } from '../code-examples/ag-grid/data-grid/data-manager/data-manager-data-grid-docs-demo.module';
 import { DataGridDemoComponent as InlineHelpDataGridDemoComponent } from '../code-examples/ag-grid/data-grid/inline-help/data-grid-demo.component';
 import { DataGridDemoModule as InlineHelpDataGridDemoModule } from '../code-examples/ag-grid/data-grid/inline-help/data-grid-demo.module';
 import { DataGridPagingDemoComponent } from '../code-examples/ag-grid/data-grid/paging/data-grid-paging-demo.component';
